move gtag to helpers file
