+++ ui/components/Analytics/index.js
@@ -2,25 +2,10 @@ import Script from 'next/script';
 import { getCookie } from 'cookies-next';
 import { useRouter } from 'next/router';
 import { useEffect } from 'react';
+import { pageview } from '../../helpers/gtag';
 
 export const GA_TRACKING_ID = process.env.NEXT_PUBLIC_TRACKING_ID;
 export const GA_TAG_ID = process.env.NEXT_PUBLIC_TAG_ID;
-
-// https://developers.google.com/analytics/devguides/collection/gtagjs/pages
-export const pageview = (url) => {
-  window.gtag('config', GA_TAG_ID, {
-    page_path: url,
-  });
-};
-
-// https://developers.google.com/analytics/devguides/collection/gtagjs/events
-export const event = ({ action, category, label, value }) => {
-  window.gtag('event', action, {
-    event_category: category,
-    event_label: label,
-    value: value,
-  });
-};
 
 export function Analytics() {
   const router = useRouter();
