+++ ui/package.json
@@ -28,7 +28,6 @@
     "node-fetch": "^3.1.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "react-ga4": "^1.4.1",
     "react-markdown": "^6.0.3",
     "urlencoded-body-parser": "^3.0.0"
   },
